refactor: address PR feedback for TCP connection time optimization
