+++ packages/transport-tcp/src/socket-to-conn.ts
@@ -81,10 +81,6 @@ export const toMultiaddrConnection = (socket: Socket, options: ToConnectionOptio
   // https://nodejs.org/dist/latest-v16.x/docs/api/net.html#socketsettimeouttimeout-callback
   socket.setTimeout(inactivityTimeout)
   
-  // Ensure TCP_NODELAY is set to true to disable Nagle's algorithm
-  // This reduces latency by sending data immediately without buffering
-  socket.setNoDelay(true)
-
   socket.once('timeout', () => {
     timedOut = true
     log('%s %s socket read timeout', direction, lOptsStr)
